+++ src/daemon-command/shutdown.ts
@@ -2,15 +2,14 @@ import colors from "colors";
 import { Socket } from "net";
 
 import { Daemon } from "../commands/start-daemon.command";
-import { killProcess } from "./kill-process";
 
 export const shutdown = async (daemon: Daemon, socket?: Socket): Promise<void> => {
     console.log(`${colors.bgGreen.bold.black(" dev-pm ")} shutting down`);
 
     await Promise.all(
         daemon.scripts.map((script) => {
-            daemon.scriptStatus[script.name] = "stopped";
-            killProcess(daemon, socket, script.name);
+            script.status = "stopped";
+            script.killProcess(socket);
         }),
     );