+++ src/daemon-command/status.daemon-command.ts
@@ -6,25 +6,18 @@ import { Daemon } from "../commands/start-daemon.command";
 import { scriptsMatchingPattern } from "./scripts-matching-pattern";
 
 export function statusDaemonCommand(daemon: Daemon, socket: Socket, scriptName: string | null /* null means all */): void {
-    const { processes } = daemon;
     const scriptsToProcess = scriptsMatchingPattern(daemon, scriptName);
-
-    const response = scriptsToProcess.map((script) => {
-        const process = processes[script.name];
-        return {
-            name: script.name,
-            running: process ? !process.killed : false,
-            pid: process && !process.killed ? process.pid : undefined,
-        };
-    });
 
     const table = new CLITable({
         head: [colors.blue.bold("Script"), colors.blue.bold("Status"), colors.bold.blue("PID")],
         colWidths: [100, 20, 20],
         style: { compact: true },
     });
-    response.forEach((item) => {
-        table.push([item.name, item.running ? colors.green("Running") : colors.red("Stopped"), item.pid?.toString()]);
+    scriptsToProcess.forEach((script) => {
+        const running = script.process ? !script.process.killed : false;
+        const status = script.status === "waiting" ? colors.yellow("Waiting") : running ? colors.green("Running") : colors.red("Stopped");
+        const pid = script.process && !script.process.killed ? script.process.pid : undefined;
+        table.push([script.name, status, pid?.toString()]);
     });
     socket.write(table.toString());
     socket.end();