+++ src/script-definition.type.ts
@@ -1,5 +1,6 @@
 export interface ScriptDefinition {
     name: string;
     script: string;
-    group: string | string[];
+    group?: string | string[];
+    waitOn?: string | string[];
 }