+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 -   stop command added that allows stopping individual processes
 -   start is not blocking anymore (a start-daemon command is now the long running proccess)
 -   pm config file isn't configurable anymore
+-   waitOn support directly in script definition
 
 # 1.0.0